Vboot Reference: Add kernel image verification benchmark.
Refactor duplicate code for the firmware image benchmark. Also fixes some functions that manipulate kernel blobs (use uint64_t instead if int).

Review URL: http://codereview.chromium.org/1079009

+++ include/kernel_image.h
@@ -106,7 +106,7 @@ uint8_t* GetKernelConfigBlob(const KernelImage* image);
  *
  * Caller owns the returned pointer and must Free() it.
  */
-uint8_t* GetKernelBlob(const KernelImage* image, int* blob_len);
+uint8_t* GetKernelBlob(const KernelImage* image, uint64_t* blob_len);
 
 /* Write kernel data from [image] to a file named [input_file].
  *
